added bitmask filtration method, appears to provide no performance boost at the moment, but it does mean filtering required & disallowed mods should be effectively free from here on

+++ src/main.rs
@@ -49,8 +49,8 @@ fn debug_prompts() {
     println!("Last combo:");
     print_combo(&combinations[count - 1]);
     
-    let required_mods: Vec<usize> = Vec::new();
-    let disallowed_mods: Vec<usize> = Vec::new();
+    let required_mods: Vec<u8> = Vec::new();
+    let disallowed_mods: Vec<u8> = Vec::new();
 
     println!("Filtering illegal pairs...");
     filter_combinations(&mut combinations, required_mods.as_slice(), disallowed_mods.as_slice());
